requirements.txt
acres==0.5.0
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
ddddocr==1.0.6
distro==1.9.0
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.20.3
flatbuffers==25.12.19
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanfriendly==10.0
idna==3.11
itsdangerous==2.2.0
jiter==0.12.0
looseversion==1.3.0
lxml==6.0.2
mpmath==1.3.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
numpy==2.4.1
onnxruntime==1.23.2
openai==2.15.0
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pathlib==1.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
playwright==1.57.0
protobuf==6.33.4
prov==2.1.1
puremagic==1.30
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
pyee==13.0.0
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdfium2==5.3.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
pyxnat==1.6.4
rdflib==7.5.0
requests==2.32.5
scipy==1.17.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
sympy==1.14.0
tqdm==4.67.1
traits==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
xlrd==2.0.2
xlsxwriter==3.2.9

